Add more array perf tests
